services/worker/pyproject.toml
[tool.poetry]
name = "worker"
version = "0.1.0"
description = "Temporal worker service"
authors = ["AI PM Team <team@example.com>"]
packages = [{ include = "worker" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
temporalio = "^1.7.0"
pydantic-settings = "^2.5.2"
opentelemetry-api = "^1.28.2"
opentelemetry-sdk = "^1.28.2"
opentelemetry-exporter-otlp = "^1.28.2"
ai-pm-common = { path = "../../libs/py/common", develop = true }

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
